# server/docs/x-HanJiang-postman-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "汉江（HanJiang）",
    "description": "一个基于 FastAPI 框架深度封装的生产级 Python Web 应用框架",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/v1/auth/login": {
      "post": {
        "tags": [
          "身份认证"
        ],
        "summary": "用户登录",
        "description": "用户名或邮箱 + 密码登录，成功后返回访问/刷新令牌",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/v1/auth/refresh": {
      "post": {
        "tags": [
          "身份认证"
        ],
        "summary": "刷新令牌",
        "description": "使用刷新令牌换取新的访问/刷新令牌对",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/v1/auth/logout": {
      "post": {
        "tags": [
          "身份认证"
        ],
        "summary": "退出登录",
        "description": "清除当前用户的 Redis 登录态，使令牌立即失效（需 Bearer 令牌）",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Username",
            "description": "用户名或邮箱"
          },
          "password": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Password",
            "description": "密码"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "description": "登录请求模型。\n\n支持用户名或邮箱 + 密码登录。\n\nAttributes:\n    username: 用户名或邮箱\n    password: 密码"
      },
...
      "RefreshTokenRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "minLength": 10,
            "title": "Refresh Token",
            "description": "刷新令牌"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshTokenRequest",
        "description": "刷新令牌请求模型。\n\nAttributes:\n    refresh_token: 刷新令牌"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}